refactor: rename vip-248 to vip-249 and add description
