+++ simulations/vip-249/simulations.ts
@@ -4,7 +4,7 @@ import { ethers } from "hardhat";
 
 import { expectEvents } from "../../src/utils";
 import { forking, testVip } from "../../src/vip-framework";
-import { COMPTROLLER, FDUSD_SUPPLY, OLD_FDUSD_SUPPLY, vFDUSD, vip248 } from "../../vips/vip-248/bscmainnet";
+import { COMPTROLLER, FDUSD_SUPPLY, OLD_FDUSD_SUPPLY, vFDUSD, vip249 } from "../../vips/vip-249/bscmainnet";
 import COMTROLLER_ABI from "./abi/comptroller.json";
 
 forking(35700072, () => {
@@ -22,7 +22,7 @@ forking(35700072, () => {
     });
   });
 
-  testVip("VIP-248 Chaos Labs Recommendations", vip248(), {
+  testVip("VIP-249 Chaos Labs Recommendations", vip249(), {
     callbackAfterExecution: async (txResponse: TransactionResponse) => {
       await expectEvents(txResponse, [COMTROLLER_ABI], ["NewSupplyCap"], [1]);
     },
